React->Django Connection
This is just a small change, but I was able to get the login functionality working. The react allows the user to interact with webpage and then uses an API request to send the email and password to the django backend for authentication. I used some AI to help change the files, so some of the code I don't fully understand yet.